requirements.txt
accelerate==1.10.1
Bio==1.8.0
biopython==1.85
bitsandbytes==0.47.0
biotite==1.4.0
datasets==4.0.0
esm==3.0.3
fake_useragent==1.5.1
gradio==6.6
gradio_molecule3d==0.0.7
joblib==1.4.2
lxml==5.3.0
matplotlib==3.10.6
seaborn==0.13.2
markdown==3.7
mdx_truly_sane_lists==1.3
pandas==2.3.2
peft==0.17.1
plotly==6.0.1
Requests==2.32.3
scikit_learn==1.6.1
scipy==1.16.1
sentencepiece==0.2.0
torchmetrics==1.8.1
tqdm==4.67.0
transformers==4.48.1
wandb==0.21.3
vplm==0.0.1
dotenv==0.9.9
langchain==0.3.27 
langchain-community==0.3.31
langchain-core==0.3.79
langchain-openai==0.3.35
langchain-text-splitters==0.3.11
langsmith==0.4.37
lightning-utilities==0.15.2
openai==2.4.0
sortedcontainers==2.4.0
